SalesLoft Client
I just want to have an entity that is in charge of handling the requests
to the SalesLoft API, without having to think about the authorization
token, converting response to a JSON and so.

Also, just created a SalesLoft Initializer to load the api-key when the
app starts. This way, I have an accessor to retrieve the API key without
having to check the ENV-vars. Also, used just "fetch", in order to make
the app blow if the key is not present, basically making it mandatory.